Friend-adding framework
The framework is in place to create relationships between friends.
There's a button that changes text depending on the relationship between
the current user and another user. We can't actually make the
relationship yet because we need to be able to edit other user's
information, which isn't allowed on the client. To be fixed.
Please reset your database so that users with the proper fields can be
created.