Client-Server breakup
Client functionality is in the /client/client.js file, and the
server-side functionality is in the /server/server.js file and is not
sent to the client.
I tried removing the "insecure" package (which I thought I had already
done), but doing so would require a lot of rethinking the code to make
sure everything works, and it's not necessary for this implementation.
Still something to keep in mind.